# [cpp] use clang++