clang++ does improve