@jinyus 2023-10-20
Main branch:
```
C++ | 36.00 ms | 519.00 ms | 4.57 s
```

This PR:
```
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
          Time (mean ± σ):     230.7 ms ±   2.1 ms    [User: 200.8 ms, System: 29.5 ms]
          Range (min … max):   227.8 ms … 234.2 ms    10 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 337 ms
        Processing time (w/o IO): 341 ms
        Processing time (w/o IO): 342 ms
          Time (mean ± σ):      1.188 s ±  0.001 s    [User: 1.049 s, System: 0.140 s]
          Range (min … max):    1.188 s …  1.189 s    2 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 2938 ms
        Processing time (w/o IO): 2936 ms
        Processing time (w/o IO): 2928 ms
          Time (mean ± σ):      5.690 s ±  0.165 s    [User: 5.081 s, System: 0.386 s]
          Range (min … max):    5.573 s …  5.806 s    2 runs

```